bridge/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-bridge"
version = "1.0.0"
description = "Aegis BFT quorum clock consumer bridge"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "protobuf>=6.31.1",
    "websockets>=12.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
]

[project.scripts]
aegis-bridge = "bridge.server_main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bridge*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"
